added Authorization class to destroy listings by their owner only

+++ App/controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Controllers;
 
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -23,7 +25,7 @@ class ListingController
 
 	public function index()
 	{
-		$listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+		$listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
 		loadView('listings/index', [
 			'listings' => $listings
@@ -88,7 +90,7 @@ class ListingController
 
 		$newListingData = array_intersect_key($_POST, array_flip($allowedFileds));
 
-		$newListingData['user_id'] = 1;
+		$newListingData['user_id'] = Session::get('user')['id'];
 
 		$newListingData = array_map('sanitize', $newListingData);
 
@@ -160,9 +162,16 @@ class ListingController
 		];
 		$listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+		// Check if listing exists
 		if (!$listing) {
 			ErrorController::notFound('Job offer not found');
 			return;
+		}
+
+		//Authorize user
+		if (!Authorization::isOwner($listing['user_id'])) {
+			$_SESSION['error_message'] = 'You are not authorized to delete this listing';
+			return redirect('/listings/' . $listing['id']);
 		}
 
 		$this->db->query('DELETE FROM listings WHERE id = :id', $params);
